requirements.txt
python-dotenv==1.0.0
loguru==0.7.3
ipykernel==6.25.1
pillow==11.3.0
ultralytics==8.3.203
pytest==8.3.5
pre-commit==4.2.0
nbstripout==0.6.1
ruff==0.11.5
mypy==1.5.0
pip-licenses>=4.5.1
pip-audit>=2.7.3
shapely==2.1.2
tqdm==4.66.4
fiftyone==0.23.6
types-Pillow==10.2.0.20240822
black==25.1.0
